fix auth function
