+++ petals/src/api/auth.ts
@@ -1,3 +1,4 @@
+import * as model from '../model';
 import * as util from './util';
 
 interface JwtCreateRequest {
@@ -5,9 +6,8 @@ interface JwtCreateRequest {
   password: string;
 }
 
-interface JwtCreateResponse {
+interface JwtCreateResponse extends model.RequestError {
   token: string;
-  non_field_errors: string[];
 }
 
 export const jwtCreate = async (username: string, password: string): Promise<JwtCreateResponse> => {
@@ -19,7 +19,6 @@ export const jwtCreate = async (username: string, password: string): Promise<Jwt
 };
 
 export const logout = () => {
-  // TODO: API
   localStorage.removeItem('token');
 };
 
@@ -27,80 +26,25 @@ interface JwtRefreshRequest {
   token: string;
 }
 
-interface JwtRefreshResponse {
+interface JwtRefreshResponse extends model.RequestError {
   token: string;
-  non_field_errors: string[];
 }
 
-const jwtRefresh = async (data: JwtRefreshRequest): Promise<JwtRefreshResponse> => {
-  const res = await util.sendRequest(util.Methods.Post, '/auth/jwt/refresh/', data, false);
-  if (res.status >= 400) {
-    throw await res.json();
-  }
-  return await res.json();
-};
-
-export const refreshToken = async (token: string): Promise<boolean> => {
-  const req: JwtRefreshRequest = { token };
-  return jwtRefresh(req)
-    .then(res => {
-      localStorage.setItem('token', res.token);
-      return true;
-    })
-    .catch(errJson => {
-      return false;
-    });
+export const jwtRefresh = async (token: string): Promise<JwtRefreshResponse> => {
+  const data: JwtRefreshRequest = { token };
+  return util.sendRequest(util.Methods.Post, '/auth/jwt/refresh/', data, false);
 };
 
 interface JwtVerifyRequest {
   token: string;
 }
 
-interface JwtVerifyResponse {
+interface JwtVerifyResponse extends model.RequestError {
   token: string;
   non_field_errors: string[];
 }
 
-const jwtVerify = async (data: JwtVerifyRequest): Promise<JwtVerifyResponse> => {
-  const res = await util.sendRequest(util.Methods.Post, '/auth/jwt/verify/', data, false);
-  if (res.status >= 400) {
-    throw await res.json();
-  }
-  return await res.json();
-};
-
-export const verifyToken = async (token: string): Promise<boolean> => {
-  const req: JwtVerifyRequest = { token };
-  return jwtVerify(req)
-    .then(res => {
-      return true;
-    })
-    .catch(errJson => {
-      return false;
-    });
-};
-
-interface JwtPayload {
-  user_id: number;
-  username: string;
-  exp: number;
-  email: string;
-}
-
-export const parseJwtTokenPayload = (token: string): JwtPayload => {
-  const base64Token = token.split('.')[1];
-  const decodedToken = new Buffer(base64Token, 'base64');
-  return JSON.parse(decodedToken.toString());
-};
-
-export const isLogin = (): boolean => {
-  const token = localStorage.getItem('token');
-  if (!token) {
-    return false;
-  }
-
-  const payload = parseJwtTokenPayload(token);
-  const now = Math.floor(new Date().getTime() / 1000);
-
-  return now <= payload.exp;
+export const jwtVerify = async (token: string): Promise<JwtVerifyResponse> => {
+  const data: JwtVerifyRequest = { token };
+  return util.sendRequest(util.Methods.Post, '/auth/jwt/verify/', data, false);
 };
